Got movement working well and some enemies.
Haven't worked on this in a few months. Was meant to be for a gamejam, but spent too much time on the movement and unnecessary features that I didn't make any levels or much else besides a few basic creatures. May come back to it later if I feel inspired.